+++ src/ent/stats.c
@@ -2,6 +2,7 @@
 
 #include <math.h>
 #include <stdbool.h>
+#include <stdint.h>
 #include <stdlib.h>
 
 #include "../util/io_util.h"
@@ -40,8 +41,9 @@ void stat_push(Stat *stat, double value, uint64_t value_num_samples) {
   stat->num_samples += value_num_samples;
   double old_mean = stat->mean;
   double value_minus_old_mean = value - old_mean;
-  stat->mean = old_mean + (((double)value_num_samples) / stat->num_samples) *
-                              value_minus_old_mean;
+  stat->mean =
+      old_mean + (((double)value_num_samples) / (double)stat->num_samples) *
+                     value_minus_old_mean;
   stat->sum_of_mean_differences_squared =
       stat->sum_of_mean_differences_squared +
       ((double)value_num_samples) * value_minus_old_mean * (value - stat->mean);
@@ -98,7 +100,7 @@ void stats_combine(Stat **stats, int number_of_stats, Stat *combined_stat) {
     combined_num_unique_samples += num_unique_samples;
     uint64_t num_samples = stat_get_num_samples(stats[i]);
     combined_num_samples += num_samples;
-    combined_mean += stat_get_mean(stats[i]) * num_samples;
+    combined_mean += stat_get_mean(stats[i]) * (double)num_samples;
   }
   if (combined_num_samples == 0) {
     combined_stat->num_unique_samples = 0;
@@ -107,12 +109,13 @@ void stats_combine(Stat **stats, int number_of_stats, Stat *combined_stat) {
     combined_stat->sum_of_mean_differences_squared = 0;
     return;
   }
-  combined_mean = combined_mean / combined_num_samples;
+  combined_mean = combined_mean / (double)combined_num_samples;
 
   double combined_error_sum_of_squares = 0;
   for (int i = 0; i < number_of_stats; i++) {
     double stdev = stat_get_stdev(stats[i]);
-    combined_error_sum_of_squares += (stdev * stdev) * get_estimator(stats[i]);
+    combined_error_sum_of_squares +=
+        (stdev * stdev) * (double)get_estimator(stats[i]);
   }
 
   double combined_sum_of_squares = 0;
@@ -120,7 +123,7 @@ void stats_combine(Stat **stats, int number_of_stats, Stat *combined_stat) {
     uint64_t num_samples = stat_get_num_samples(stats[i]);
     double mean = stat_get_mean(stats[i]);
     double mean_diff = (mean - combined_mean);
-    combined_sum_of_squares += (mean_diff * mean_diff) * num_samples;
+    combined_sum_of_squares += (mean_diff * mean_diff) * (double)num_samples;
   }
   double combined_sum_of_mean_differences_squared =
       combined_sum_of_squares + combined_error_sum_of_squares;
