+++ src/util/math_util.c
@@ -15,7 +15,7 @@ double z_to_p_cdf(double z) { return (0.5 * (1 + erf(z / sqrt(2.0)))) * 100; }
 // Assumes that a continuity correction from binomial to normal distribution
 // has already been applied
 double odds_that_player_is_better(double corrected_sampled_win_pct,
-                                  int total_games) {
+                                  uint64_t total_games) {
   return z_to_p_cdf((corrected_sampled_win_pct - 0.5) * 2 *
                     sqrt((double)total_games));
 }
